CHORE: move report modal into a separate cog
